+++ modules/graphql-codegen/src/main/scala/anduin/graphql/codegen/SourceFile.scala
@@ -1,5 +1,5 @@
 // Copyright (C) 2017 Anduin Transactions, Inc.
 
-package anduin.graphql.codegen2
+package anduin.graphql.codegen
 
 final case class SourceFile(path: String) extends AnyVal
