--- modules/graphql-codegen/src/main/scala/anduin/graphql/codegen2/package.scala
@@ -1,7 +1,0 @@
-// Copyright (C) 2017 Anduin Transactions, Inc.
-
-package anduin.graphql
-
-package object codegen2 {
-  type Result[A] = Either[CodegenException, A]
-}
